feat: pluralize names with helper method
Had an issue with models whose names already ends in 's'.
While it makes sense in the app to either change the model names
or implement name_plural directly,
it also helps with the initial app usage if the model names
are just a little cleaner earlier on. :)

As noted, the inflect package may be more useful there.
I leave that to the maintainer's discretion.