Fix test passing in model type instead of instance
The type signature of get_save_redirect_url specifies that it must
receive an instance of a `ModelView` but this test is passing in a model
type.

Fixing this in preperation for multiple primary key support. The URL
generating functions will become more complicated meaning they must
receive an actual model instance as specified in its function signature.