fix: remove request.url usage in templates
request.url is a handy tool to reach for to get the current URL,
but it depends heavily on the configuration of the application.
This can return the incorrect URL or incorrect scheme
when used behind a proxy, for instance.

Instead of going to route of implementing middleware to correct it,
take advantage of default browser behavior.
An empty URL used for 'href' or 'action'
is automatically filled with the full URL from the browser's context.
This also allows relative URLs from the current location,
including URLs that require query params
(these can include other jinja variables so long as we 'urlencode' them for safety).