Add comment explaining if statement
This if statement returns true if the function is the same as ModelView.
It will return false if the user has overridden the function in their
own class.

The if statement surrounds the relationships for loop since, if they
have overridden the function to filter relationships, we don't want to
add all of the relationships back in afterwards.