Mock request object
Making reference to the `batman.id` value for the primary key in the
request object was causing PostgreSQL tests to hang. This commit matches
the way that the request object is being mocked in the
`test_models_actions` file. Local tests with a PostgreSQL instance built
in Docker are now working.