fix: search_query with multiple related objects
For example there is a model Cat with a foreign key to the User and with a foreign key to the Address,
and the Address also has a foreign key to the User.

In this case Select(Cat).join(User).join(Address) produces such query:
"... JOIN users ON user.id = cats.owner_id JOIN addresses ON users.id = addresses.user_id"
i.e. search will be performed by owner address, not by cat address. The same is true for join_from method.

This commit fixes it by using a model attribute as a parameter for the join.